HienThiDSThanhToan

+++ database/migrations/2025_05_21_152315_remove_total_from_payments_table.php
@@ -5,12 +5,16 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
-{public function up()
+{
+    public function up()
 {
     Schema::table('payments', function (Blueprint $table) {
-        $table->dropColumn('total');
+        if (Schema::hasColumn('payments', 'total')) {
+            $table->dropColumn('total');
+        }
     });
 }
+
 
 public function down()
 {
